finalize styling system, fix & test controls and fill the example with all the controls
we went back to an inherited style system as this allows for custom controls to have their own fields.
also this should allow us to animate between two styles in the future.

we did lose the ability to change elements of the next control's style on the fly.
but that system had many problems to do with extensibility, animation, performance and API complication.